# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is the MyMovieDB API definition",
    "version": "1.0.0",
    "title": "MyMovieDB",
    "contact": {
      "email": "jeremythen16@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
    "/movies": {
      "get": {
        "tags": ["movies"],
        "summary": "Get all movies",
        "operationId": "getMovies",
        "description": "Get all the movies. Anybody, logged in or not can call this endpoint.",
        "security": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved movies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMoviesResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnkownError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["movies"],
        "summary": "Creates a movie",
        "description": "Create a movie. Only the admin can access this endpoint.",
        "operationId": "createMovie",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer authorization token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer jwttoken"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieCreationPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created movie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddMovieResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter",
            "content": {
              "appliation/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMovieDbBadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}": {
      "get": {
        "tags": ["movies"],
        "summary": "Get a movie by id",
        "description": "Get a movie by the specified id.",
        "operationId": "getMovieById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The movie id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved movie specified by that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMovieByIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter",
            "content": {
              "appliation/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMovieDbBadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}/reviews": {
      "get": {
        "tags": ["movies"],
        "summary": "Get the reviews for this movie",
        "operationId": "getReviewsByMovieId",
        "description": "Get a review by the specified reviewId",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The movie id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved movie's reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetMovieReviewsResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Etc",
            "content": {
              "appliation/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevieBadResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["movies"],
        "summary": "Add a review to a movie for the logged in user",
        "description": "Add a review to a movie specified by it's id. Any logged in user can access this endpoint. A user can only add a review to one specific movie. If the user tries to add a review to the same movie, the previous review will be overriden by the new review. The logged in user will be the owner of this review.",
        "operationId": "addMovieReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The movie id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer authorization token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer jwttoken"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddReviewPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created movie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddReviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter",
            "content": {
              "appliation/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMovieDbBadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}/casts": {
      "get": {
        "tags": ["movies"],
        "summary": "Get a list of the casts in this movie",
        "operationId": "getMovieCasts",
        "description": "Get a list of the casts in this movie",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The movie id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved movie's casts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddCastResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Etc",
            "content": {
              "appliation/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevieBadResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["movies"],
        "summary": "Add a cast to a movie by the movie id",
        "description": "Add an cast to a movie by the movie id",
        "operationId": "addMovieCast",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The movie id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer authorization token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer jwttoken"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCastPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created movie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddCastResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter",
            "content": {
              "appliation/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMovieDbBadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}/directors": {
      "get": {
        "tags": ["movies"],
        "summary": "Get a list of the directors in this movie",
        "operationId": "getMovieDirectors",
        "description": "Get a list of the directors in this movie",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The movie id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved movie's directors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddDirectorResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Etc",
            "content": {
              "appliation/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevieBadResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["movies"],
        "summary": "Add a director to a movie by the movie id",
        "description": "Add an director to a movie by the movie id",
        "operationId": "addMovieDictor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The movie id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer authorization token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer jwttoken"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDirectorPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created movie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddDirectorResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter",
            "content": {
              "appliation/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyMovieDbBadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}/disable": {
      "put": {
        "tags": ["movies"],
        "summary": "Diactivates the movie",
        "operationId": "diactivateMovie",
        "description": "Diactivates the movie specified by the id. After this, the movie will not be visible. Only an admin can access this endpoint",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The movie id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer authorization token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer jwttoken"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated movie",
            "content": {
              "appliation/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisableMovieResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter",
            "content": {
              "appliation/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisableMovieBadResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/movies/pagination/offset/{offset}": {
      "get": {
        "tags": ["movies"],
        "summary": "Get all movies after offset count movie record",
        "operationId": "getPaginatedMoviesOffset",
        "description": "Get all the movies that are after the movie with the offset position",
        "security": [],
        "parameters": [
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "description": "The offset from where to get movies",
            "example": 5,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved movies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetMoviesResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnkownError"
                }
              }
            }
          }
        }
      }
    },
    "/movies/pagination/limit/{limit}": {
      "get": {
        "tags": ["movies"],
        "summary": "Get movies up to limit count",
        "operationId": "getPaginatedMoviesLimit",
        "description": "Get movies up to limit count. If limit is 5, then up to 5 movies will be returned",
        "security": [],
        "parameters": [
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "description": "The count limit to return",
            "example": 2,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved movies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetMoviesResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnkownError"
                }
              }
            }
          }
        }
      }
    },
    "/movies/pagination/offset/{offset}/limit/{limit}": {
      "get": {
        "tags": ["movies"],
        "summary": "Get movies after th especified offset and up to limit count",
        "operationId": "getPaginatedMoviesOffsetAndLimit",
        "description": "Get movies up to limit count. If limit is 5, then up to 5 movies will be returned, after the movie with the offset position",
        "security": [],
        "parameters": [
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "description": "The offset from where to get movies",
            "example": 5,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "description": "The count limit to return",
            "example": 3,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved movies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetMoviesResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad input parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnkownError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "MyMovieDbBadResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The data that was requested. null if any issue requesting it",
            "type": "string",
            "example": null
          },
          "success": {
            "description": "The requested operation was successful",
            "type": "boolean",
            "example": false
          },
          "errorCode": {
            "description": "An error code of an error that happened while requesting the data",
            "type": "string",
            "example": "MOVIE_INVALID_PAYLOAD"
          },
          "errorMessages": {
            "description": "The detailed message of the error that happened",
            "type": "array",
            "items": {
              "example": "No movie review payload was provided"
            }
          }
        }
      },
...
      "AddReviewPayload": {
        "type": "object",
        "required": ["movieId", "reviewerStars"],
        "properties": {
          "movieId": {
            "type": "integer",
            "example": 5
          },
          "reviewerStars": {
            "description": "The stars that the reviewer user is giving to this movie",
            "type": "integer",
            "example": 5,
            "default": 5
          },
          "comment": {
            "type": "string",
            "description": "The reviewer's or user's comment for this movie review",
            "example": "I'm voting 5 to this movie because it's awesome"
          }
        }
      },
...
      "MovieCreationPayload": {
        "type": "object",
        "required": ["title", "year", "time"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Avengers End Game",
            "default": "Avengers End Game"
          },
          "year": {
            "description": "The year of th emovie",
            "type": "integer",
            "example": 1,
            "default": 1
          },
          "time": {
            "description": "The duration of the movie in minutes",
            "type": "string",
            "example": 120
          },
          "language": {
            "description": "The language the movie is originally in",
            "type": "string",
            "example": "English"
          },
          "distributor": {
            "description": "The distributor of the movie",
            "type": "string",
            "example": "Universal Pictures"
          }
        }
      },
...
      "RevieBadResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The data that was requested. null if any issue requesting it",
            "type": "string",
            "example": null
          },
          "success": {
            "description": "The requested operation was successful",
            "type": "boolean",
            "example": false
          },
          "errorCode": {
            "description": "An error code of an error that happened while requesting the data",
            "type": "string",
            "example": "MOVIE_NOT_FOUND"
          },
          "errorMessages": {
            "description": "The detailed message of the error that happened",
            "type": "array",
            "items": {
              "example": "Movie with the specified id was not found"
            }
          }
        }
      },
      "DisableMovieResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The data that was requested. null if any issue requesting it",
            "type": "string",
            "example": {
              "disabled": true
            }
          },
          "success": {
            "description": "The requested operation was successful",
            "type": "boolean",
            "example": true
          },
          "errorCode": {
            "description": "An error code of an error that happened while requesting the data",
            "type": "string",
            "example": null
          },
          "errorMessages": {
            "description": "The detailed message of the error that happened",
            "type": "array",
            "items": {
              "example": ""
            }
          }
        }
      },
...
      "UnkownError": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The data that was requested. null if any issue requesting it",
            "type": "string",
            "example": null
          },
          "success": {
            "description": "The requested operation was successful",
            "type": "boolean",
            "example": false
          },
          "errorCode": {
            "description": "An error code of an error that happened while requesting the data",
            "type": "string",
            "example": "UNKNOWN_ERROR"
          },
          "errorMessages": {
            "description": "Detailed messages of the errors that happened",
            "type": "array",
            "items": {
              "example": "There was an error with your request. Try again later."
            }
          }
        }
      },
      "AddReviewResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The data that was requested. null if any issue requesting it",
            "type": "object",
            "properties": {
              "review": {
                "$ref": "#/components/schemas/Review"
              }
            }
          },
          "success": {
            "description": "The requested operation was successful",
            "type": "boolean",
            "example": true
          },
          "errorCode": {
            "description": "An error code of an error that happened while requesting the data",
            "type": "string",
            "example": ""
          },
          "errorMessages": {
            "description": "Detailed messages of the errors that happened",
            "type": "array",
            "items": {
              "example": ""
            }
          }
        }
      },
      "GetMoviesResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The data that was requested. null if any issue requesting it",
            "type": "object",
            "properties": {
              "movies": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "success": {
            "description": "The requested operation was successful",
            "type": "boolean",
            "example": true
          },
          "errorCode": {
            "description": "An error code of an error that happened while requesting the data",
            "type": "string",
            "example": ""
          },
          "errorMessages": {
            "description": "Detailed messages of the errors that happened",
            "type": "array",
            "items": {
              "example": ""
            }
          }
        }
      },
      "AddMovieResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The data that was requested. null if any issue requesting it",
            "type": "object",
            "properties": {
              "movie": {
                "type": "object",
                "$ref": "#/components/schemas/NewMovieCreated"
              }
            }
          },
          "success": {
            "description": "The requested operation was successful",
            "type": "boolean",
            "example": true
          },
          "errorCode": {
            "description": "An error code of an error that happened while requesting the data",
            "type": "string",
            "example": ""
          },
          "errorMessages": {
            "description": "Detailed messages of the errors that happened",
            "type": "array",
            "items": {
              "example": ""
            }
          }
        }
      },
      "GetMovieByIdResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The data that was requested. null if any issue requesting it",
            "type": "object",
            "properties": {
              "movie": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          },
          "success": {
            "description": "The requested operation was successful",
            "type": "boolean",
            "example": true
          },
          "errorCode": {
            "description": "An error code of an error that happened while requesting the data",
            "type": "string",
            "example": ""
          },
          "errorMessages": {
            "description": "Detailed messages of the errors that happened",
            "type": "array",
            "items": {
              "example": ""
            }
          }
        }
      },
      "GetMovieReviewsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The data that was requested",
            "type": "object",
            "properties": {
              "review": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "success": {
            "description": "The requested operation was successful",
            "type": "boolean",
            "example": true
          },
          "errorCode": {
            "description": "An error code of an error that happened while requesting the data",
            "type": "string",
            "example": ""
          },
          "errorMessages": {
            "description": "Detailed messages of the errors that happened",
            "type": "array",
            "items": {
              "example": ""
            }
          }
        }
      },
...
      "AddCastPayload": {
        "type": "object",
        "properties": {
          "actorId": {
            "type": "integer",
            "example": 5
          },
          "role": {
            "type": "string",
            "example": "Hulk"
          }
        }
      },
      "AddCastResponsePayload": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The id of this association movie_actor record",
            "example": 1
          },
          "actorId": {
            "type": "integer",
            "example": 5
          },
          "role": {
            "type": "string",
            "example": "Hulk"
          }
        }
      },
      "AddDirectorPayload": {
        "type": "object",
        "properties": {
          "directorId": {
            "type": "integer",
            "example": 5
          }
        }
      },
...
      "AddCastResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The data that was requested",
            "type": "object",
            "properties": {
              "cast": {
                "$ref": "#/components/schemas/AddCastResponsePayload"
              }
            }
          },
          "success": {
            "description": "The requested operation was successful",
            "type": "boolean",
            "example": true
          },
          "errorCode": {
            "description": "An error code of an error that happened while requesting the data",
            "type": "string",
            "example": ""
          },
          "errorMessages": {
            "description": "Detailed messages of the errors that happened",
            "type": "array",
            "items": {
              "example": ""
            }
          }
        }
      },
      "AddDirectorResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The data that was requested",
            "type": "object",
            "properties": {
              "directors": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AddDirectorResponsePayload"
                }
              }
            }
          },
          "success": {
            "description": "The requested operation was successful",
            "type": "boolean",
            "example": true
          },
          "errorCode": {
            "description": "An error code of an error that happened while requesting the data",
            "type": "string",
            "example": ""
          },
          "errorMessages": {
            "description": "Detailed messages of the errors that happened",
            "type": "array",
            "items": {
              "example": ""
            }
          }
        }
      }
    }
  }
}